[Add]録音音声ファイル化機能追加／音声合成機能トライアル追加
