+++ app/assets/config/manifest.js
@@ -1,2 +1,3 @@
 //= link_tree ../images
+//= link_tree ../audios
 //= link_directory ../stylesheets .css
